resources/labs/python_opengl/requirements.txt
PyOpenGL~=3.1.5
PyOpenGL_accelerate~=3.1.5
